src/requirements.txt
fastapi
black
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
python-dotenv
aiofiles
python-multipart
pillow
pytesseract